src/parsers/cobol.py
"""COBOL language parser using ANTLR."""

import antlr4
from typing import List
from .base import LanguageParser, ASTNode, NodeType
# from .antlr.cobol import COBOLLexer, COBOLParser as AntlrCOBOLParser  # Generated from ANTLR grammar

class COBOLParser(LanguageParser):
    """Parser for COBOL language."""
    
    def __init__(self):
        self.lexer = None
        self.parser = None
    
    def parse_file(self, filepath: str) -> ASTNode:
        """Parse COBOL source file."""
        with open(filepath, 'r') as f:
            source = f.read()
        return self.parse_string(source)
    
    def parse_string(self, source_code: str) -> ASTNode:
        """Parse COBOL source code string."""
        # Initialize ANTLR lexer and parser
        input_stream = antlr4.InputStream(source_code)
        # self.lexer = COBOLLexer(input_stream)
        # token_stream = antlr4.CommonTokenStream(self.lexer)
        # self.parser = AntlrCOBOLParser(token_stream)
        
        # Parse and build AST
        # parse_tree = self.parser.compilationUnit()
        # ast = self._build_ast(parse_tree)
        
        # return ast
        raise NotImplementedError("ANTLR parser not yet generated. Uncomment lines after generating.")
    
    def _build_ast(self, parse_tree) -> ASTNode:
        """Convert ANTLR parse tree to our AST format."""
        # Implementation: Walk ANTLR tree and construct ASTNode objects
        # This is language-specific logic
        pass
    
    def extract_functions(self, ast: ASTNode) -> List[ASTNode]:
        """Extract COBOL paragraphs/procedures as functions."""
        functions = []
        # Walk AST and find PROCEDURE DIVISION elements
        self._find_nodes_by_type(ast, NodeType.PROCEDURE, functions)
        return functions
    
    def extract_variables(self, ast: ASTNode) -> List[ASTNode]:
        """Extract COBOL DATA DIVISION variables."""
        variables = []
        # Walk AST and find working-storage variables
        self._find_nodes_by_type(ast, NodeType.VARIABLE, variables)
        return variables
    
    def get_dependencies(self, ast: ASTNode) -> List[str]:
        """Extract COPY statements and CALL targets."""
        dependencies = []
        # Find all COPY and CALL statements
        return dependencies
    
    def _find_nodes_by_type(self, node: ASTNode, target_type: NodeType, 
                           results: List[ASTNode]) -> None:
        """Recursively find nodes of specific type."""
        if node.node_type == target_type:
            results.append(node)
        for child in node.children:
            self._find_nodes_by_type(child, target_type, results)
