pyproject.toml
[project]
name = "relian"
version = "0.1.0"
description = "Universal Legacy Refactoring Substrate"
authors = [{name = "Aldrich Khaalis Wooden, Sr.", email = "akwooden@zuup.io"}]
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "langchain>=0.1.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "xgboost>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "neo4j>=5.14.0",
    "psycopg2-binary>=2.9.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "pylint>=3.0.0",
    "mypy>=1.7.0",
    "antlr4-python3-runtime>=4.13.0"
]

[project.optional-dependencies]
dev = [
    "ipython>=8.18.0",
    "jupyter>=1.0.0",
    "pre-commit>=3.5.0"
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term"
